refactor: 토스터로 이름 변경
