+++ linkmind/src/main/java/com/app/toaster/domain/Reminder.java
@@ -1,4 +1,4 @@
-package com.app.linkmind.domain;
+package com.app.toaster.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -7,42 +7,31 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
 @Entity
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Toast {
+public class Reminder {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long Id;
+	Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "toast_id")
 	private User user;
 
-	private String title;
-
-	private String linkUrl;
-
-	private Boolean isRead;
+	@OneToOne
+	@JoinColumn(name = "category")
+	private Category category;
 
 	@Builder
-	public Toast(User user, String title, String linkUrl) {
+	public Reminder(User user, Category category, String title) {
 		this.user = user;
-		this.title = title;
-		this.linkUrl = linkUrl;
-		this.isRead = false;
+		this.category = category;
 	}
 
-	public void updateTitle(String title) {
-		this.title = title;
-	}
-
-	public void updateIsRead(Boolean isRead) {
-		this.isRead = isRead;
-	}
 }
