+++ linkmind/src/main/java/com/app/toaster/domain/Toast.java
@@ -1,4 +1,4 @@
-package com.app.linkmind.domain;
+package com.app.toaster.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -7,31 +7,42 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity
 @Getter
+@Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Reminder {
+public class Toast {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	Long id;
+	private Long Id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "toast_id")
 	private User user;
 
-	@OneToOne
-	@JoinColumn(name = "category")
-	private Category category;
+	private String title;
+
+	private String linkUrl;
+
+	private Boolean isRead;
 
 	@Builder
-	public Reminder(User user, Category category, String title) {
+	public Toast(User user, String title, String linkUrl) {
 		this.user = user;
-		this.category = category;
+		this.title = title;
+		this.linkUrl = linkUrl;
+		this.isRead = false;
 	}
 
+	public void updateTitle(String title) {
+		this.title = title;
+	}
+
+	public void updateIsRead(Boolean isRead) {
+		this.isRead = isRead;
+	}
 }
