+++ linkmind/src/main/java/com/app/toaster/domain/User.java
@@ -1,4 +1,4 @@
-package com.app.linkmind.domain;
+package com.app.toaster.domain;
 
 import java.util.Objects;
